+++ Assets/XCharts/Editor/Utility/ChartEditorHelper.cs
@@ -117,8 +117,9 @@ public class ChartEditorHelper
 
         float defaultWidth = drawRect.width;
         float defaultX = drawRect.x;
-        drawRect.width = EditorGUIUtility.labelWidth - EditorGUI.indentLevel * 15;
+        drawRect.width = EditorGUIUtility.labelWidth - 5;
         var displayName = string.IsNullOrEmpty(moduleName) ? prop.displayName : moduleName;
+
         toggle = EditorGUI.Foldout(drawRect, toggle, displayName, bold ? foldoutStyle : EditorStyles.foldout);
         if (moduleToggle[key] != toggle)
         {